Drop board revision from BOARD-CFG names
The number of BOARD-CFGs is growing considerably. However the differences
between different board-revsions are only minor. So privide the differences
as subnodes in the BOARD-CFG file itself and drop the board revision from
the BOARD-CFG names.

So instead of having PCoreMX8MM-FERT3.110.dts, PCoreMX8MM-FERT3.120.dts and
PCoreMX8MM-FERT3.130.dts, there is only one file PCoreMX8MM-FERT3.dts now
and this file contains three subnodes for these three board revisions:

board-cfg {
	/* Common board features */
	...

	/* Board-revision specific features */
	PCoreMX8MM-FERT3.110 {
		board-rev = <110>;
		//have-rtc-pcf85063;
		//have-rtc-pcf85263;
	};
	PCoreMX8MM-FERT3.120 {
		board-rev = <120>;
		have-rtc-pcf85063;
		//have-rtc-pcf85263;
	};
	PCoreMX8MM-FERT3.130 {
		board-rev = <130>;
		//have-rtc-pcf85063;
		have-rtc-pcf85263;
	};
};

The subnodes contain those features that differ in these board revisions.
All other features, that are the same in all board revisions, are still
defined in the parent board-cfg node.

- Add function fs_image_get_board_rev_subnode() that determines the
  appropriate subnode for the current BOARD-ID (given in compare-id).
  The compare-id has to be set to the BOARD-ID before calling this
  function, e.g. by calling fs_image_set_compare_id() or
  fs_image_set_board_id_from_cfg().
- Add function fs_image_get_board_rev() to return only the board-rev
  of the BOARD-ID (set in compare-id).
- Add function_fs_image_get_board_rev_subnode_f(). This is similar to
  fs_image_get_board_rev_subnode(), but can be called in f-phase of
  U-Boot when no variables are available. In this case, we cannot use
  compare-id and have to determine the BOARD-ID directly from the
  BOARD-CFG in OCRAM. This function also returns the board-rev from
  the BOARD-ID via an additional argument of the function.
- Add functions fs_image_getprop() and fs_image_getprop_u32(). They
  behave like fdt_getprop() and fdt_getprop_u32_default_node(), but they
  check the board-revision subnode first, and if the property is not
  found, also the parent board-cfg node. Use these functions wherever the
  board-cfg node is queried. (Remark: queries to nboot-info can still use
  the original function as there are no such subnodes.)
- Drop the board-revision from all BOARD-CFG files and add board revision
  subnodes where appropriate. The board-rev property is now part of these
  subnodes again and can not be set automatically in nboot-info.dts by the
  Makefile anymore.

+++ include/fs_cfg_info.h
@@ -19,7 +19,7 @@
 struct cfg_info {
 	enum boot_device boot_dev;
 	u8 board_type;
-	u16 board_rev;
+	u32 board_rev;
 	u32 features;
 	u32 dram_size;
 	u32 dram_chips;
